==> 5_rename_files.R <==
file.copy('Images/density_obs.png', 'Paper Objects/fig_1.png', TRUE)
file.copy('Images/prs_main.png', 'Paper Objects/fig_2.png', TRUE)
file.copy('Images/heat_all_bmi_raw_prs_k.png', 'Paper Objects/fig_3.png', TRUE)
file.copy('Images/both_sep_ridit.png', 'Paper Objects/fig_4.png', TRUE)
file.copy('Images/prs_3.png', 'Paper Objects/fig_5.png', TRUE)

file.copy('Images/mean_prs_sep.png', 'Paper Objects/fig_s1.png', TRUE)
file.copy('Images/prs_all.png', 'Paper Objects/fig_s2.png', TRUE)
file.copy('Images/prs_all_dep.png', 'Paper Objects/fig_s3.png', TRUE)
file.copy('Images/quant_all_bmi_raw_prs_k.png', 'Paper Objects/fig_s4.png', TRUE)
file.copy('Images/quant_all_bmi_raw_prs_r_adult.png', 'Paper Objects/fig_s5.png', TRUE)
file.copy('Images/quant_all_bmi_raw_prs_v.png', 'Paper Objects/fig_s6.png', TRUE)
file.copy('Images/quant_all_bmi_raw_prs_r_child.png', 'Paper Objects/fig_s7.png', TRUE)
file.copy('Images/both_all_ridit.png', 'Paper Objects/fig_s8.png', TRUE)
file.copy('Images/sep_additive_r2.png', 'Paper Objects/fig_s9.png', TRUE)
file.copy('Images/mult_all_bmi_raw.png', 'Paper Objects/fig_s10.png', TRUE)
file.copy('Images/prs_sex.png', 'Paper Objects/fig_s11.png', TRUE)
file.copy('Images/prs_corr.png', 'Paper Objects/fig_s12.png', TRUE)
file.copy('Images/prs_all_corrected.png', 'Paper Objects/fig_s13.png', TRUE)
file.copy('Images/attrit_means.png', 'Paper Objects/fig_s14.png', TRUE)
file.copy('Images/attrit_prs_k_all_bmi_raw.png', 'Paper Objects/fig_s15.png', TRUE)
file.copy('Images/attrit_prs_r_adult_all_bmi_raw.png', 'Paper Objects/fig_s16.png', TRUE)
file.copy('Images/attrit_prs_v_all_bmi_raw.png', 'Paper Objects/fig_s17.png', TRUE)
file.copy('Images/attrit_prs_r_child_all_bmi_raw.png', 'Paper Objects/fig_s18.png', TRUE)
file.copy('Images/attrit_r2_prs_k.png', 'Paper Objects/fig_s19.png', TRUE)
file.copy('Images/attrit_r2_prs_r_adult.png', 'Paper Objects/fig_s20.png', TRUE)
file.copy('Images/attrit_r2_prs_v.png', 'Paper Objects/fig_s21.png', TRUE)
file.copy('Images/attrit_r2_prs_r_child.png', 'Paper Objects/fig_s22.png', TRUE)
file.copy('Images/bmi_by_prs_missing.png', 'Paper Objects/fig_s23.png', TRUE)
